The user can now add songs to an existing playlist
I am highly aware that it is bad practice of what I'm doing when it comes to version management however I have gone many days with code that simply doesn't work and I refuse for the life of me to push non functional code onto the repository. Plus on top of that I wasn't sure if my github pro subscription woud run out, and I'm not willing to put this onto a public repository for privacy reasons.