The user can now log in and register
Bootstrap however decided to snort crack and overdose on my end so I have honest to god no idea why the style on the page decided to stop working, maybe it'll work for someone else but I am NOT fixing it because I reverted to a previous version, reinstalled bootstrap several times in every way possible, tried upgrading and downgrading and even tried switching to the web version but none of these solutions worked. I would love to see the confused look on my teachers face when they take on the mystery of why bootstrap just shit the bed here.